test(shared-server): consolidate client-state shell coverage
